Seperating Timeline into Timeline and Animation
