+++ fury/animation/__init__.py
@@ -1,0 +1,2 @@
+from fury.animation.animation import Animation
+from fury.animation.timeline import Timeline
